change in components,refactor to little jsx  files, easy to work now
