+++ client/src/components/admin/Admin.jsx
@@ -1,38 +1,18 @@
-import React, { useState } from "react";
-import { useLogin } from "../../hooks/useAdminLogin";
-import { ToastContainer } from "react-toastify";
-import "react-toastify/dist/ReactToastify.css";
+import React from "react";
+import { Header } from "../layout/header/Header.jsx";
+import { Footer } from "../layout/footer/Footer.jsx";
+import { AdminMessages } from "./AdminMessages.jsx";
+import { AdminReseservations } from "./AdminReservation.jsx";
 import "./Admin.css";
+import "./AdminReservation.css";
 
-export const Admin = () => {
-  const [username, setUsername] = useState("");
-  const [password, setPassword] = useState("");
-
-  const { loginHandler } = useLogin();
-
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-    await loginHandler(username, password);
-  };
+export const AllReservations = () => {
   return (
-    <div>
-      <form className="admin-form" onSubmit={handleSubmit}>
-        <input
-          type="text"
-          placeholder="Korisnicko ime"
-          value={username}
-          onChange={(e) => setUsername(e.target.value)}
-        />
-
-        <input
-          type="password"
-          placeholder="Lozinka"
-          value={password}
-          onChange={(e) => setPassword(e.target.value)}
-        />
-        <button type="submit">Uloguj se</button>
-      </form>
-      <ToastContainer />
-    </div>
+    <>
+      <Header />
+      <AdminReseservations />
+      <AdminMessages />
+      <Footer />
+    </>
   );
 };
